+++ tests/test_data/edge_cases/34_doc404_return_and_yield_double_unwrap/google.py
@@ -1,0 +1,11 @@
+from typing import Any, Dict, Iterator
+
+
+def f() -> Iterator[Dict[str, Any]]:
+    """Do work.
+
+    Yields:
+        Dict[str, Any]: The outputs.
+    """
+    yield {}
+    return